Move invite folder into on-boarding
