+++ apps/dashboard/src/app/onboarding/invite/invite.component.scss
@@ -1,4 +1,4 @@
-@import 'apps/dashboard/src/variables';
+@import '../../../variables';
 
 .setup-invite {
   display: flex;
