+++ apps/dashboard/src/app/onboarding/invite/invite.component.ts
